+++ CHANGELOG.md
@@ -28,6 +28,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - ai2/general-cirrascale
     ```
     See the docs for the `BeakerExecutor` for more information on the input parameters.
+- **Step class:**
+  - Added `workspace_metadata` argument and field to the step class API.
+- **Weights & Biases integration:**
+  - You can now change the artifact kind for step result artifacts by adding a field
+    called "artifact_kind" to a step's ``workspace_metadata`` dictionary.
+    For models, setting "artifact_kind" to "model" will add the corresponding artifact to W&B's new model zoo.
 
 ### Changed